pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vpn-guard-agent"
version = "0.1.0"
description = "CLI anti-abuse monitor for VPN/proxy servers: Xray/3x-ui, WireGuard/AmneziaWG, OpenVPN"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "VPN Guard Agent contributors"}]
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pyyaml>=6.0.1",
]

[project.scripts]
vpn-guard = "vpn_guard.main:app"

[tool.setuptools.packages.find]
include = ["vpn_guard*"]
